fix clear all
